ml/pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "lightgbm>=4.6.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "shap>=0.51.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
